requirements.txt
pandas
numpy
scikit-learn
nltk
streamlits